nexo_dev/nexo/requirements.txt
asgiref==3.8.1
anyio==3.6.2
argon2-cffi==21.3.0
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
Django==5.1.5
django-allauth==65.4.1
django-crispy-forms==2.3
django-google-sso==8.0.0
django-template-debug==0.3.5
djangorestframework==3.15.1
crispy-bootstrap5==0.7
et_xmlfile==2.0.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
loguru==0.7.3
mysqlclient==2.2.7
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.6.0
win32_setctime==1.2.0
